+++ crates/wdl-ast/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+#### Fixed
+
+* If a WDL >=v1.2 workflow has a non-boolean value for the "allow nested inputs" hint, the metadata will also be
+    searched for `allowNestedInputs: <bool>` instead of prematurely dis-allowing nested inputs
+    ([#445](https://github.com/stjude-rust-labs/wdl/pull/445)).
+
 ## 0.12.1 - 05-02-2025
 
 _A patch bump was required because an error was made during the release of `wdl` v0.13.0 regarding dependencies._